requirements.txt
asgiref==3.3.1
cycler==0.10.0
Django==3.1.7
django-appconf==1.0.4
django-imagekit==2.0.4
djangorestframework==3.12.4
et-xmlfile==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.1
numpy==1.20.2
openpyxl==3.0.7
pandas==1.2.3
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1